# [WIP] Reduce lock contention in chunk write queue