--- a/tsdb/chunks/chunk_write_queue_test.go
+++ b/tsdb/chunks/chunk_write_queue_test.go
@@ -50,7 +50,7 @@ func TestChunkWriteQueue_GettingChunkFromQueue(t *testing.T) {
 	require.NoError(t, q.addJob(job))
 
 	// Retrieve chunk from the queue.
-	gotChunk := q.get(ref)
+	gotChunk := q.get(ref, 0)
 	require.Equal(t, testChunk, gotChunk)
 }
 
@@ -243,7 +243,7 @@ func BenchmarkChunkWriteQueue_addJob(b *testing.B) {
 					go func() {
 						for range issueReadSignal {
 							// We don't care about the ID we're getting, we just want to grab the lock.
-							_ = q.get(ChunkDiskMapperRef(0))
+							_ = q.get(ChunkDiskMapperRef(0), 0)
 						}
 					}()
 
@@ -364,7 +364,7 @@ func benchmarkChunkWriteQueue_lockContentionOnChunkRefMap(b *testing.B, readConc
 
 			for jobID := startPos; jobID < startPos+concurrencyAdjustedBN; jobID++ {
 				jobIDWrapped := jobID % queueSize
-				q.get(ChunkDiskMapperRef(jobIDWrapped))
+				q.get(ChunkDiskMapperRef(jobIDWrapped), HeadSeriesRef(jobIDWrapped))
 			}
 		}(readerID)
 	}