This is a prototype where I wanted to test whether we could reduce the lock contention on `chunkWriteQueue.chunkRefMapMtx` by sharding the map which it protects. In this prototype I divided it into `1024`, which is an arbitrary number and likely subject to change.

The first commit adds a benchmark which tries to measure the read performance while constantly writing through the queue.
The second commit implements the sharding of the `chunkRefMapMtx`.

This is the benchstat comparison of the benchmark results with/without the second commit:
```
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_1/with_queueSize_1000-4         1.17µs ± 0%    0.05µs ± 0%  -95.59%
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_1/with_queueSize_1000000-4      2.22µs ± 0%    0.22µs ± 0%  -90.08%
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_10/with_queueSize_1000-4         139ns ± 0%      21ns ± 0%  -84.55%
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_10/with_queueSize_1000000-4      271ns ± 0%      60ns ± 0%  -77.70%
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_100/with_queueSize_1000-4        112ns ± 0%      19ns ± 0%  -82.64%
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_100/with_queueSize_1000000-4     142ns ± 0%      58ns ± 0%  -59.02%
```

TBH this looks a bit too good to be true, @codesome if you get a chance it would be great if you could also take a look at these changes to check if they make sense.

Related: https://github.com/prometheus/prometheus/issues/10377