@codesome 2022-03-03
@prymitive would you be willing to help test this out and let us know if this improves the rule evaluation and scrape durations? Thanks!