@codesome 2022-03-03
16384 is the shard size used for sharding the series map currently in the Head. So if that is sane enough, we can keep that as default.

At a scale of 6M active series we saw this queue to reduce latency and not increase it. I am little surprised that the latency increased with the queue here because without the queue, your reads will fight with a synchronous write of chunks while this queue makes the write asynchronous and reduce the peak latency of reads.

I haven't looked at your profiles from the issue yet, will take a look at them next week. If that can show what is being slow after holding the lock, we may be able to improve it further.