@roidelapluie 2022-03-03
> Is reverting chunk write queue an option? From what I understand it was suppose to improve latency for remote read/write and this scalability issue feels like too much of a sacrifice.

For me this would definitively be an option.