@replay 2022-03-03
> Is reverting chunk write queue an option? From what I understand it was suppose to improve latency for remote read/write and this scalability issue feels like too much of a sacrifice.
> 
> For me this would definitively be an option.

The reason why we added it is because due to the remote write spikes that we saw without the queue we couldn't scale to the size which we needed, with the queue we were able to scale further. However, I think that our cluster which I'm talking about likely had less concurrent reads than the one @prymitive is running, which would explain why we didn't see the read latency issue.

I would prefer to not completely remove the chunk write queue, but we could make it possible to disable it completely by setting the queue size to `0`.