@roidelapluie 2022-03-03
So far,  in the Prometheus server, remote write has always been 'async' and having delays there has been acceptable. Delays in queries are a lot less acceptable, Prometheus should compute recording rules in time to be able to send timely alerts.