@replay 2022-03-03
> We might want to have multiple writer routines. Currently only one shard is locked by the writer at a time.
All readers start at 0 currently, hence the bias of writer only blocking one shard at a time will be amplified here. We might want to start all readers at different jobID.
With 1, we want to have a higher write concurrency than read. 1x, 10x, 100x and maybe even 1000x if we can run it sanely on a laptop.

I have updated the benchmark according to your suggestions. I force pushed to move the benchmark improvements before the commits which implement the sharded `chunkRefMap`, to make it easy to compare the benchmark results before/after sharding.

Those are the results which I get with `-test.benchtime=10s`:
```
name                                                                                                              old time/op    new time/op    delta                 
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_10/write_concurrency_100/with_queueSize_1000-4         40.0ns ± 0%    16.3ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_10/write_concurrency_100/with_queueSize_1000000-4      62.8ns ± 0%    43.1ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_10/write_concurrency_1000/with_queueSize_1000-4        42.6ns ± 0%    15.9ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_10/write_concurrency_1000/with_queueSize_1000000-4     84.0ns ± 0%    46.1ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_100/write_concurrency_100/with_queueSize_1000-4        33.5ns ± 0%    12.8ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_100/write_concurrency_100/with_queueSize_1000000-4     43.9ns ± 0%    32.3ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_100/write_concurrency_1000/with_queueSize_1000-4       46.5ns ± 0%    10.4ns ± 0%   ~     (p=1.000 n=1+1)
ChunkWriteQueue_lockContentionOnChunkRefMap/read_concurrency_100/write_concurrency_1000/with_queueSize_1000000-4    46.6ns ± 0%    31.3ns ± 0%   ~     (p=1.000 n=1+1)
```