@replay 2022-03-03
> We might want to have multiple writer routines. Currently only one shard is locked by the writer at a time.

Btw I just realized that even with many writer routines, only one shard will ever be locked by the writers at a time, because `c.isRunningMtx` serializes the access to `c.chunkRefMapMtx`: https://github.com/prometheus/prometheus/blob/c726072e98dfa84eafef9bc02ea8cd175a43221e/tsdb/chunks/chunk_write_queue.go#L140

Another shard may be locked by `processJob()` at the same time, but that's also only running in a single thread.

TBH this makes it harder for me to understand why reading chunks would be slowed down so significantly by the writes, if there are never more than two shards locked and there can't ever be a long queue of writers waiting for the `RWMutex` then acquiring the read lock *should* not be slow...