@codesome 2022-03-07
> One thing to note is that although initially the rate of gets is pretty small (30 per second) it later grows a lot to a few thousands. Here's a better screenshot taken from another Prometheus instance that scrapes affected one (affected instance gets gaps in recorded metrics due to this issue):

Yup that is expected. Because, with majority of series with same scrape interval, their chunks will get full and get m-mapped within the same period of time.