@bboreham 2024-01-16
Discussed in the bug-scrub meeting; we think a lot of changes have since been made in Prometheus around the issue this was meant to fix, so will close to clean up the backlog.
If any of these ideas in this PR are worth revisiting, please feel free to reopen or create a new PR.